refactor(blorktools): css3d animation util
